requirements.txt
Django==1.4
South==0.7.5
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.2.1
django-heroku-memcacheify==0.3
django-pylibmc-sasl==0.2.4
psycopg2==2.4.5
pylibmc==1.2.3
python-ptrace==0.6.4
requests==0.13.2
wsgiref==0.1.2
-e git+git://github.com/toastdriven/django-haystack.git#egg=django-haystack
-e git+git://github.com/caseypt/pyelasticsearch.git#egg=pyelasticsearch
gunicorn==0.14.5